+++ src/org/netbeans/modules/php/laravel/LaravelPhpModuleCustomizerExtender.java
@@ -4,6 +4,8 @@ Licensed to the Apache Software Foundation (ASF)
 package org.netbeans.modules.php.laravel;
 
 import java.util.EnumSet;
+import java.util.HashMap;
+import java.util.Map;
 import javax.swing.JComponent;
 import javax.swing.event.ChangeListener;
 import org.netbeans.modules.php.api.phpmodule.PhpModule;
@@ -12,6 +14,7 @@ import org.netbeans.modules.php.laravel.ui.customizer.LaravelCustomizerPanel;
 import org.netbeans.modules.php.spi.framework.PhpModuleCustomizerExtender;
 import org.openide.util.HelpCtx;
 import org.netbeans.modules.php.laravel.project.ComposerPackages;
+import org.openide.filesystems.FileObject;
 import org.openide.util.NbBundle.Messages;
 
 /**
@@ -22,7 +25,6 @@ public class LaravelPhpModuleCustomizerExtender extends PhpModuleCustomizerExten
 
     private final PhpModule phpModule;
     private final ComposerPackages composerPackages;
-    
     private final boolean isFrameworkEnabledOnProject;
 
     // @GuardedBy(EDT)
@@ -30,7 +32,7 @@ public class LaravelPhpModuleCustomizerExtender extends PhpModuleCustomizerExten
     
     LaravelPhpModuleCustomizerExtender(PhpModule phpModule) {
         this.phpModule = phpModule;
-        composerPackages = ComposerPackages.fromPhpModule(phpModule);
+        composerPackages = ComposerPackages.fromProjectDir(phpModule.getProjectDirectory());
         isFrameworkEnabledOnProject = LaravelPreferences.hasEnabledConfigured(phpModule);
     }