+++ src/org/netbeans/modules/php/laravel/commands/ArtisanCommandSupport.java
@@ -4,8 +4,14 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.WeakHashMap;
+import org.netbeans.api.project.Project;
 import org.netbeans.modules.php.api.phpmodule.PhpModule;
+import static org.netbeans.modules.php.laravel.PhpNbConsts.NB_PHP_PROJECT_TYPE;
+import static org.netbeans.modules.php.laravel.project.ComposerPackages.fromProjectDir;
 import org.netbeans.modules.php.spi.framework.commands.FrameworkCommand;
+import org.netbeans.spi.project.LookupProvider;
+import org.openide.util.Lookup;
+import org.openide.util.lookup.Lookups;
 
 /**
  *
@@ -13,22 +19,7 @@ import org.netbeans.modules.php.spi.framework.commands.FrameworkCommand;
  */
 public class ArtisanCommandSupport {
 
-    private static final Map<String, ArtisanCommandSupport> INSTANCES = new WeakHashMap<>();
     public List<FrameworkCommand> commands = new ArrayList<>();
-
-    public static ArtisanCommandSupport getInstance(PhpModule phpModule)
-    {
-        String projectPath = phpModule.getProjectDirectory().getPath();
-
-        synchronized (INSTANCES) {
-            ArtisanCommandSupport artisanCommandSupport = INSTANCES.get(projectPath);
-            if (artisanCommandSupport == null) {
-                artisanCommandSupport = new ArtisanCommandSupport();
-                INSTANCES.put(projectPath, artisanCommandSupport);
-            }
-            return artisanCommandSupport;
-        }
-    }
 
     private ArtisanCommandSupport() {
 
@@ -36,5 +27,26 @@ public class ArtisanCommandSupport {
     
     public List<FrameworkCommand> getCommands(){
         return commands;
+    }
+    
+    public void addCommand(ArtisanCommand command) {
+        commands.add(command);
+    }
+    
+    public static ArtisanCommandSupport fromPhpModule(PhpModule phpModule) {
+        Project project = phpModule.getLookup().lookup(Project.class);
+        return project.getLookup().lookup(ArtisanCommandSupport.class);
+    }
+    
+    @LookupProvider.Registration(projectType = NB_PHP_PROJECT_TYPE)
+    public static LookupProvider createJavaBaseProvider() {
+        return new LookupProvider() {
+            @Override
+            public Lookup createAdditionalLookup(Lookup baseContext) {
+                return Lookups.fixed(
+                    new ArtisanCommandSupport()
+                );
+            }
+        };
     }
 }