+++ src/org/netbeans/modules/php/laravel/editor/HyperlinkProviderImpl.java
@@ -208,7 +208,7 @@ public class HyperlinkProviderImpl implements HyperlinkProviderExt {
                         }
                         break;
                 }
-
+                break;
             case ALT_HYPERLINK:
                 JTextComponent focused = EditorRegistry.focusedComponent();
                 if (focused != null && focused.getDocument() == doc) {