+++ src/org/netbeans/modules/php/laravel/project/ProjectUtils.java
@@ -4,6 +4,7 @@ import javax.swing.text.Document;
 import org.netbeans.api.project.Project;
 import org.netbeans.modules.editor.NbEditorUtilities;
 import org.netbeans.modules.php.api.phpmodule.PhpModule;
+import org.netbeans.modules.php.laravel.LaravelPhpFrameworkProvider;
 import org.netbeans.modules.php.laravel.preferences.LaravelPreferences;
 import org.netbeans.spi.project.ui.support.ProjectConvertors;
 import org.openide.filesystems.FileObject;
@@ -17,6 +18,11 @@ public class ProjectUtils {
     public static Project get(Document doc) {
         FileObject file = NbEditorUtilities.getFileObject(doc);
         Project projectOwner = ProjectConvertors.getNonConvertorOwner(file);
+        return projectOwner;
+    }
+    
+    public static Project get(FileObject fo) {
+        Project projectOwner = ProjectConvertors.getNonConvertorOwner(fo);
         return projectOwner;
     }
 
@@ -33,9 +39,20 @@ public class ProjectUtils {
     public static boolean isInLaravelModule(PhpModule module) {
         return LaravelPreferences.isEnabled(module);
     }
-
+    
     public static PhpModule getPhpModule(Document doc){
         Project pr = get(doc);
+        if (pr == null) {
+            return null;
+        }
+        return pr.getLookup().lookup(PhpModule.class);
+    }
+
+    public static PhpModule getPhpModule(FileObject fo){
+        Project pr = get(fo);
+        if (pr == null) {
+            return null;
+        }
         return pr.getLookup().lookup(PhpModule.class);
     }
 }